Load correct stack memory if thread record is wrong (#1)
Some minidumps contain invalid thread memory descriptors pointing to memory
regions other than the ones referred to by the RSP. This is the case in UE4, for
example. This causes the stackwalker to terminate early as it cannot load the
actual stack memory.

Other Minidump processors such as WinDbg or the Visual Studio debugger can still
process such minidumps correctly. They do not load the stack memory region based
on the information obtained from the TEB and instead just follow the stack
walker.

This patch adds the same behavior to the minidump processor. If the thread
memory region does not cover the stack pointer, the region is replaced with the
one covering it, if any.